Implement modular arithmetic interpreter
- Refactor arithmetic interpreter tests into a common behavior.
- Moves the representation function into abstracting.
